# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API E-Commerce",
        "description": "Documentación de los endpoints del E-Commerce",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/orders/{order}/payments": {
            "post": {
                "tags": [
                    "Pagos"
                ],
                "summary": "Procesar un pago con Stripe para una orden",
                "description": "Crea un PaymentIntent en Stripe, ejecuta el cobro y registra la transacción localmente.",
                "operationId": "storePayment",
                "parameters": [
                    {
                        "name": "order",
                        "in": "path",
                        "description": "ID de la orden a pagar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos requeridos para procesar el pago",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "payment_method_id"
                                ],
                                "properties": {
                                    "payment_method_id": {
                                        "description": "ID del método de pago generado por Stripe (ej. pm_card_visa en testing)",
                                        "type": "string",
                                        "example": "pm_card_visa"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pago procesado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "payment": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "order_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "provider": {
                                                    "type": "string",
                                                    "example": "stripe"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "succeeded"
                                                },
                                                "amount": {
                                                    "type": "integer",
                                                    "example": 2000
                                                },
                                                "currency": {
                                                    "type": "string",
                                                    "example": "usd"
                                                },
                                                "stripe_payment_intent_id": {
                                                    "type": "string",
                                                    "example": "pi_3MtwBwLkd901xA0Z12345678"
                                                },
                                                "paid_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2026-09-13T20:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "client_secret": {
                                            "type": "string",
                                            "example": "pi_3MtwBwLkd901xA0Z12345678_secret_ABC123"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado (La orden no pertenece al usuario autenticado)"
                    },
                    "422": {
                        "description": "Error de validación en la solicitud (Falta payment_method_id)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}